Pull ws hooks into own file
